Use row number instead of line number in errors
The DictReader.line_num value is not necessarily the same as the row
number as a single row can cover multiple lines. The row number is
a more useful value in error messages as it is expected users will
typically look at the CSV files using a tool like Excel.